+++ src/stolas/struct/__init__.py
@@ -1,6 +1,7 @@
 """S: Struct module."""
 
+from .replace import replace
 from .struct import struct
 from .trait import MissingImplementationWarning, trait
 
-__all__ = ["struct", "trait", "MissingImplementationWarning"]
+__all__ = ["struct", "trait", "MissingImplementationWarning", "replace"]